Fix various Eclipse warnings
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/sanselan/trunk@995630 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/sanselan/formats/png/chunks/PNGChunkgAMA.java
@@ -36,7 +36,7 @@ public class PNGChunkgAMA extends PNGChunk
 
 	public double getGamma()
 	{
-		return 1.0 / ((double) Gamma / 100000.0);
+		return 1.0 / (Gamma / 100000.0);
 	}
 
 }
